move auxpow to separate module
